+++ components/shared/custome/c-input-field.tsx
@@ -7,25 +7,31 @@ export default function CInputField({
   className,
   classNameParent,
   label,
-  iconSvg,
+  iconSvg: IconSvg,
   iconImg,
-  iconSize,
+  iconSize = 20,
   field,
   ...props
 }: IFieldInput) {
-  const isError = field && field.state.meta.errors.length;
+  const isError = field && !!field.state.meta.errors.length;
+  const hasIcon = !!IconSvg || !!iconImg;
   return (
     <div className={cn("w-full", classNameParent)}>
       {label && <Label>{label}</Label>}
-      <div className={cn("relative", field && "mb-[25px]")}>
-        <div className="absolute top-1/2 left-3 -translate-y-1/2">
-          {iconSvg && iconSvg}
+      <div className={cn("relative")}>
+        <div
+          className={cn(
+            "absolute top-1/2 left-3 -translate-y-1/2",
+            !props.value && "*:grayscale *:text-gray-400",
+          )}
+        >
+          {IconSvg && <IconSvg size={iconSize} />}
           {iconImg && (
             <Image
               src={iconImg}
               alt="icon"
-              width={iconSize || 20}
-              height={iconSize || 20}
+              width={iconSize}
+              height={iconSize}
               className="shrink-0"
             />
           )}
@@ -33,20 +39,16 @@ export default function CInputField({
         <Input
           placeholder={props.placeholder}
           className={cn(className, {
-            "pl-9": iconSvg,
-            "pl-10": iconImg,
+            "!pl-10": hasIcon,
           })}
           onWheel={(event) => event.currentTarget.blur()}
           checked={props.checked}
-          isError={isError}
           {...props}
         />
-        {isError ? (
-          <em className="absolute left-0 bottom-[-20px] text-xs text-red-500">
-            {field.state.meta.errors?.map((err: any) => err.message)?.[0]}
-          </em>
-        ) : null}
       </div>
+      {isError && (
+        <em>{field.state.meta.errors?.map((err: any) => err.message)?.[0]}</em>
+      )}
     </div>
   );
 }